+++ src/pages/courses/list.tsx
@@ -1,3 +1,4 @@
+import { useContext } from "react";
 import {
   List,
   Table,
@@ -15,7 +16,11 @@ import {
 } from "@pankod/refine";
 import { ICourse, ISchoolYear } from "interfaces";
 
+import { SchoolYearContext } from "context/SchoolYearContext";
+
 export const CourseList: React.FC<IResourceComponentsProps> = () => {
+  const { schoolYearContext } = useContext(SchoolYearContext);
+
   const { tableProps, sorter } = useTable<ICourse>({
     initialSorter: [
       {
@@ -30,7 +35,7 @@ export const CourseList: React.FC<IResourceComponentsProps> = () => {
       {
           field: "school_year_id",
           operator: "eq",
-          value: localStorage.getItem("selectedSchoolYearId"),
+          value: schoolYearContext.id,
       },
     ],
   });
@@ -41,8 +46,8 @@ export const CourseList: React.FC<IResourceComponentsProps> = () => {
     optionValue: "id",
   });
 
-  console.log('getting selected school year id: ', localStorage.getItem("selectedSchoolYearId"));
-  console.log(tableProps);
+  //console.log('getting selected school year id: ', localStorage.getItem("selectedSchoolYearId"));
+  //console.log(tableProps);
 
   return (
     <List>
